requirements.txt
# --- Web Framework & API ---
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# --- Database & ORM ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# asyncpg  <-- Async Database Connection

# --- Data Analysis & Processing ---
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # Export format Arrow/Parquet

# --- Financial Data Sources ---
ccxt>=4.0.0      # Crypto (Binance)
yfinance>=0.2.0  # Stock (Yahoo Finance)

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0

# --- Development & Testing (CI/CD) ---
pytest>=7.4.0        # For running unit tests
pytest-cov>=4.1.0    # For measuring code coverage
mypy>=1.8.0          # For static type checking (Important for strict typing)
flake8>=7.0.0        # For linting (Syntax & Style check)
black>=24.0.0        # For auto-formatting code
isort>=5.13.0        # For sorting imports automatically
httpx>=0.27.0        # For testing Async APIs

# --- News & Date Handling ---
gnews>=0.3.6           # Google News RSS Scraper
python-dateutil>=2.8.2 # Robust date parsing